Comment
This will allow the developers of the server to simply know exactly which method they need to call before they even type anything. This is because comments are sent directly into the developers brain so that they can interpret and understand the code that is set before them. The power that comments hold knows no bounds and is only held back by developers who do not comment anything. They have created a rift in the understanding of code and must change their ways before the implosion of the universe.