unfortunately fixed pineapple
no longer do we have have the magnificent pineapple overtaking the users screen in game. It now is covered by the grids and buttons. This is quite a sad change, however it is necessary since the requirements require a working game instead of the most majestic picture of a pineapple.